Api and NetWorkModule

+++ app/src/main/java/com/example/bookstore/data/model/Books.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Books (
 
-	@SerializedName("isbn") val isbn : Int,
+	@SerializedName("isbn") val isbn : Long,
 	@SerializedName("title") val title : String,
 	@SerializedName("author") val author : String,
 	@SerializedName("description") val description : String,
