+++ app/src/main/java/com/example/bookstore/data/model/Results.kt
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName
 data class Results (
 
 	@SerializedName("books") val books : List<Books>,
-	@SerializedName("best_sellers") val best_sellers : List<Int>
+	@SerializedName("best_sellers") val best_sellers : List<Long>
 )
